+++ server/src/main/java/net/oneandone/stool/server/StageExistsException.java
@@ -1,0 +1,4 @@
+package net.oneandone.stool.server;
+
+public class StageExistsException extends Exception {
+}